+++ tests/test_recruiters.py
@@ -1121,50 +1121,56 @@ class TestMultiRecruiter(object):
     def test_open_recruitment(self, recruiter):
         result = recruiter.open_recruitment(n=3)
         assert len(result["items"]) == 3
-        assert result["items"][0].startswith("http://0.0.0.0:5000/ad?recruiter=cli")
-        assert result["items"][1].startswith("http://0.0.0.0:5000/ad?recruiter=cli")
-        assert result["items"][2].startswith("http://0.0.0.0:5000/ad?recruiter=hotair")
+        assert result["items"][0].startswith("http://localhost:5000/ad?recruiter=cli")
+        assert result["items"][1].startswith("http://localhost:5000/ad?recruiter=cli")
+        assert result["items"][2].startswith(
+            "http://localhost:5000/ad?recruiter=hotair"
+        )
 
     def test_open_recruitment_over_recruit(self, recruiter):
         result = recruiter.open_recruitment(n=5)
         assert len(result["items"]) == 3
-        assert result["items"][0].startswith("http://0.0.0.0:5000/ad?recruiter=cli")
-        assert result["items"][1].startswith("http://0.0.0.0:5000/ad?recruiter=cli")
-        assert result["items"][2].startswith("http://0.0.0.0:5000/ad?recruiter=hotair")
+        assert result["items"][0].startswith("http://localhost:5000/ad?recruiter=cli")
+        assert result["items"][1].startswith("http://localhost:5000/ad?recruiter=cli")
+        assert result["items"][2].startswith(
+            "http://localhost:5000/ad?recruiter=hotair"
+        )
 
     def test_open_recruitment_twice(self, recruiter):
         result = recruiter.open_recruitment(n=1)
         assert len(result["items"]) == 1
-        assert result["items"][0].startswith("http://0.0.0.0:5000/ad?recruiter=cli")
+        assert result["items"][0].startswith("http://localhost:5000/ad?recruiter=cli")
 
         result2 = recruiter.open_recruitment(n=3)
         assert len(result2["items"]) == 2
-        assert result2["items"][0].startswith("http://0.0.0.0:5000/ad?recruiter=cli")
-        assert result2["items"][1].startswith("http://0.0.0.0:5000/ad?recruiter=hotair")
+        assert result2["items"][0].startswith("http://localhost:5000/ad?recruiter=cli")
+        assert result2["items"][1].startswith(
+            "http://localhost:5000/ad?recruiter=hotair"
+        )
 
     def test_recruit(self, recruiter):
         result = recruiter.recruit(n=3)
         assert len(result) == 3
-        assert result[0].startswith("http://0.0.0.0:5000/ad?recruiter=cli")
-        assert result[1].startswith("http://0.0.0.0:5000/ad?recruiter=cli")
-        assert result[2].startswith("http://0.0.0.0:5000/ad?recruiter=hotair")
+        assert result[0].startswith("http://localhost:5000/ad?recruiter=cli")
+        assert result[1].startswith("http://localhost:5000/ad?recruiter=cli")
+        assert result[2].startswith("http://localhost:5000/ad?recruiter=hotair")
 
     def test_over_recruit(self, recruiter):
         result = recruiter.recruit(n=5)
         assert len(result) == 3
-        assert result[0].startswith("http://0.0.0.0:5000/ad?recruiter=cli")
-        assert result[1].startswith("http://0.0.0.0:5000/ad?recruiter=cli")
-        assert result[2].startswith("http://0.0.0.0:5000/ad?recruiter=hotair")
+        assert result[0].startswith("http://localhost:5000/ad?recruiter=cli")
+        assert result[1].startswith("http://localhost:5000/ad?recruiter=cli")
+        assert result[2].startswith("http://localhost:5000/ad?recruiter=hotair")
 
     def test_recruit_partial(self, recruiter):
         result = recruiter.open_recruitment(n=1)
         assert len(result["items"]) == 1
-        assert result["items"][0].startswith("http://0.0.0.0:5000/ad?recruiter=cli")
+        assert result["items"][0].startswith("http://localhost:5000/ad?recruiter=cli")
 
         result2 = recruiter.recruit(n=3)
         assert len(result2) == 2
-        assert result2[0].startswith("http://0.0.0.0:5000/ad?recruiter=cli")
-        assert result2[1].startswith("http://0.0.0.0:5000/ad?recruiter=hotair")
+        assert result2[0].startswith("http://localhost:5000/ad?recruiter=cli")
+        assert result2[1].startswith("http://localhost:5000/ad?recruiter=hotair")
 
         result3 = recruiter.recruit(n=2)
         assert len(result3) == 0
@@ -1176,14 +1182,14 @@ class TestMultiRecruiter(object):
         recruiter = MultiRecruiter()
         result = recruiter.recruit(n=10)
         assert len(result) == 8
-        assert result[0].startswith("http://0.0.0.0:5000/ad?recruiter=cli")
-        assert result[1].startswith("http://0.0.0.0:5000/ad?recruiter=cli")
-        assert result[2].startswith("http://0.0.0.0:5000/ad?recruiter=hotair")
-        assert result[3].startswith("http://0.0.0.0:5000/ad?recruiter=cli")
-        assert result[4].startswith("http://0.0.0.0:5000/ad?recruiter=cli")
-        assert result[5].startswith("http://0.0.0.0:5000/ad?recruiter=cli")
-        assert result[6].startswith("http://0.0.0.0:5000/ad?recruiter=hotair")
-        assert result[7].startswith("http://0.0.0.0:5000/ad?recruiter=hotair")
+        assert result[0].startswith("http://localhost:5000/ad?recruiter=cli")
+        assert result[1].startswith("http://localhost:5000/ad?recruiter=cli")
+        assert result[2].startswith("http://localhost:5000/ad?recruiter=hotair")
+        assert result[3].startswith("http://localhost:5000/ad?recruiter=cli")
+        assert result[4].startswith("http://localhost:5000/ad?recruiter=cli")
+        assert result[5].startswith("http://localhost:5000/ad?recruiter=cli")
+        assert result[6].startswith("http://localhost:5000/ad?recruiter=hotair")
+        assert result[7].startswith("http://localhost:5000/ad?recruiter=hotair")
 
     def test_close_recruitment(self, recruiter):
         patch1 = mock.patch("dallinger.recruiters.CLIRecruiter.close_recruitment")