+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Change Log
 
 ## [v-master](https://github.com/dallinger/dallinger/tree/master) (xxxx-xx-xx)
+- Bugfix: Fixes for Dashboard monitor layout and color issues.
 - New customizable database dashboard for viewing live experiment data.
+- Use localhost as hostname when running in debug mode by default.
 
 ## [v-6.3.1](https://github.com/dallinger/dallinger/tree/6.3.1) (2020-07-21)
 - Bugfix: Dashboard authentication now works with multiple web processes and dynos